refactor: split copilot_task_submit.rs into sub-modules
Split the 1,238-line copilot_task_submit.rs (1,022 impl) into four
sub-modules under src/copilot_task_submit/, each ≤400 impl lines:

- types.rs (278 lines): constants, CopilotSubmitOutcome, CopilotSubmitReport,
  CopilotSubmitRun, CopilotSubmitFlowAsset, TranscriptCheckpointScan and
  related scanning types
- transcript.rs (364 impl + 217 test lines): scan_transcript, classify_startup,
  classify_submit, fragment processing, and all existing tests
- orchestration.rs (357 lines): observe_startup, observe_submit,
  persist_report, build_evidence_records, finalize_session,
  ensure_copilot_submit_is_launchable
- mod.rs (88 lines): run_copilot_submit entry point and pub(crate) re-exports

All 656 lib tests continue to pass. Public API paths preserved:
crate::copilot_task_submit::{CopilotSubmitRun, CopilotSubmitReport, run_copilot_submit}
